Fix Gateway plugin zip not including version information
