+++ buildSrc/src/main/kotlin/toolkit-publishing-conventions.gradle.kts
@@ -1,11 +1,33 @@
 // Copyright 2024 Amazon.com, Inc. or its affiliates. All Rights Reserved.
 // SPDX-License-Identifier: Apache-2.0
 import org.jetbrains.intellij.platform.gradle.IntelliJPlatformType
+import org.jetbrains.intellij.platform.gradle.tasks.BuildPluginTask
+import org.jetbrains.intellij.platform.gradle.tasks.PatchPluginXmlTask
 import org.jetbrains.intellij.platform.gradle.tasks.VerifyPluginTask
+import software.aws.toolkits.gradle.buildMetadata
 import software.aws.toolkits.gradle.intellij.IdeVersions
+import software.aws.toolkits.gradle.isCi
 
 plugins {
     id("org.jetbrains.intellij.platform")
+}
+
+val ideProfile = IdeVersions.ideProfile(project)
+val toolkitVersion: String by project
+
+// please check changelog generation logic if this format is changed
+version = "$toolkitVersion-${ideProfile.shortName}"
+
+// attach the current commit hash on local builds
+if (!project.isCi()) {
+    val buildMetadata = buildMetadata()
+    tasks.withType<PatchPluginXmlTask>().configureEach {
+        pluginVersion.set("${project.version}+$buildMetadata")
+    }
+
+    tasks.named<BuildPluginTask>("buildPlugin") {
+        archiveClassifier.set(buildMetadata)
+    }
 }
 
 intellijPlatform {
