+++ plugins/amazonq/build.gradle.kts
@@ -37,22 +37,6 @@ dependencies {
     implementation(project(":plugin-amazonq:shared"))
 }
 
-val moduleOnlyJar = tasks.create<Jar>("moduleOnlyJar") {
-    archiveClassifier.set("module-only")
-    // empty jar
-}
-
-val moduleOnlyJars by configurations.creating {
-    isCanBeConsumed = true
-    isCanBeResolved = false
-    // If you want this configuration to share the same dependencies, otherwise omit this line
-    extendsFrom(configurations["implementation"], configurations["runtimeOnly"])
-}
-
-artifacts {
-    add("moduleOnlyJars", moduleOnlyJar)
-}
-
 tasks.check {
     val serviceSubdirs = project(":plugin-amazonq").subprojects
     serviceSubdirs.forEach { serviceSubDir ->
